requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-filter>=23.2
Pillow>=10.0.0  # For image handling
python-dotenv>=1.0.0
djangorestframework-simplejwt>=5.3.0  # For JWT authentication
drf-spectacular>=0.26.0  # For API documentation 